Fix progress outputs, redirect to stderr
It is standard POSIX convention to write diagnostics to stderr rather
than stdout, to avoid cluttering stdout with unnecessary information.
This is particularly try for the pagination queries of products, as
there are potentially hundreds of them.

Redirecting to stderr allows the user to configure what to do with
errors.