Fix overfetching of products
We now use a HashSet when retrieving products to ensure we don't have
duplicate products fetched from the categories.

This requires us to implement Eq and PartialEq on Product, with the
guarantee that PartialEq is a equivalence relation. Because we defined
our equivalence relation in such a way that products are only equal if
their `sku` values are equal, we satisfy all three properties of an
equivalence relation (reflexive, symmetric, and transitive).

This closes #4, and so we increase the verbosity of a failed to get
product ID for an sku to a `warn` (from `trace`).