Reduce concurrent requests
Countdown API limits requests, so reducing this value from 6 -> 2
reduces the chance of bad API responses.

In the future, we should implement retry, and also not crash the first
time we fail to collect a price. Our scraper should be resilient and
continue to scrape, reporting the failures at the end.